requirements.txt
langchain-core>=1.2.0
langchain-classic>=1.0.0
langchain-community>=0.4.0
langchain-chroma>=1.1.0
langchain-voyageai>=0.3.0
langchain-anthropic>=1.3.0
langgraph>=1.0.0
google-api-python-client>=2.100.0
youtube-transcript-api>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
voyageai>=0.3.0
aiohttp>=3.9.0
aiolimiter>=1.0.0
